requirements.txt
click~=8.1.6
Flask~=2.3.2
requests~=2.31.0
oauthlib~=3.2.2
python-dotenv~=1.0.0
SQLAlchemy==2.0.20
psycopg2-binary~=2.9.7
Flask-Login~=0.6.2
pyOpenSSL~=23.2.0